:lock: Fix security vulnerability #2
