+++ frontend/functions/urlFunctions.ts
@@ -1,3 +1,3 @@
-export default function displayPdf(url: string) {
-  return `/api/pdf/${encodeURIComponent(url)}`;
+export default function displayPdf(pathName: string) {
+  return `/api/pdf/${encodeURIComponent(pathName)}`;
 }
